http://ftp.ksu.edu.tw/FTP/FreeBSD/distfiles/xscreensaver-4.23.tar.gz -rw-r--r-- 1 zblaxell zblaxell 4939070 Oct 22  2005 xscreensaver-4.23.tar.gz b7d3f0b4a94d879a2450a268ae941b92ca4fc39e  xscreensaver-4.23.tar.gz

+++ hacks/spotlight.c
@@ -181,9 +181,9 @@ onestep (Display *dpy, Window window, Bool first_p)
     now = currentTimeInMs() + off;
 
     /* find new x,y */
-    x = ((1 + sin(((float)now) / X_PERIOD * 2. * M_PI))/2.0) 
+    x = ((1 + sin(((double)now) / X_PERIOD * 2. * M_PI))/2.0) 
       * (sizex - s/2) -s/4  + MINX;
-    y = ((1 + sin(((float)now) / Y_PERIOD * 2. * M_PI))/2.0) 
+    y = ((1 + sin(((double)now) / Y_PERIOD * 2. * M_PI))/2.0) 
       * (sizey - s/2) -s/4  + MINY;
     
     if (!first_p)
